TryFIX Top10MethodsError 0.2

+++ src/main/java/com/urlshortenerh2/dto/LinkCountsDTO.java
@@ -7,5 +7,5 @@ import lombok.Setter;
 @Setter
 public class LinkCountsDTO {
 
-    private Integer linkCounts;
+    private Long linkCounts;
 }
